Create README.md
Dear,
VISHNUPRIYA S
We would like to congratulate you on being selected for the “ Java Programming” virtual
internship position with “CodSoft”. We at CodSoft are excited that you will join our team.
The duration of the internship will be of 4 weeks, starting from 01 February 2025 to 28 February
2025. The internship is an educational opportunity for you hence the primary focus is on learning
and developing new skills and gaining hands-on knowledge. We believe that you will perform all
your tasks/projects.
As an intern, we expect you to perform all assigned tasks to the best of your ability and follow any
lawful and reasonable instructions provided to you.
We are confident that this internship will be a valuable experience for you, we look forward to
working with you and helping you achieve your career goals.
By accepting this offer, you commit to executing assigned tasks diligently and ensuring excellence in
all aspects of your work.
Best of Luck!
Thank You!